requirements.txt
# Core dependencies
langgraph>=0.2.0
langchain>=0.3.0
langchain-openai>=0.2.0
langchain-anthropic>=0.2.0
langchain-groq>=0.1.0
pydantic>=2.0.0
fastapi>=0.115.0
uvicorn>=0.30.0

# Search and extraction libraries
httpx>=0.27.0
beautifulsoup4>=4.12.0
trafilatura>=1.12.0

# Testing frameworks
pytest>=8.0.0
hypothesis>=6.100.0
pytest-mock>=3.14.0
pytest-asyncio>=0.23.0

# Additional utilities
python-dotenv>=1.0.0
networkx>=3.0
